Deduplicate inline `fn constrain` to `src/cdef.rs`.
